backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
google-generativeai==0.3.2
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
